add create_with_partition_expr_persists_manifest
